requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
antlr4-python3-runtime==4.9.3
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
certifi==2022.6.15
charset-normalizer==2.1.0
colorama==0.4.5
datasets==2.3.2
dill==0.3.5.1
filelock==3.7.1
frozenlist==1.3.0
fsspec==2022.5.0
google-pasta==0.2.0
grpcio==1.34.1
huggingface-hub==0.8.1
hydra-core==1.2.0
idna==3.3
importlib-resources==5.8.0
Keras-Preprocessing==1.1.2
multidict==6.0.2
multiprocess==0.70.13
numpy==1.19.5
omegaconf==2.2.2
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
protobuf==4.21.2
pyarrow==8.0.0
pycodestyle==2.8.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
regex==2022.6.2
requests==2.28.1
responses==0.18.0
six==1.15.0
tensorflow-estimator==2.5.0
tokenizers==0.12.1
toml==0.10.2
torch==1.12.0
tqdm==4.64.0
transformers==4.20.1
typing-extensions==4.3.0
urllib3==1.26.9
wrapt==1.12.1
xxhash==3.0.0
yarl==1.7.2
zipp==3.8.0
